Google dip fix
