+++ core/payments/src/google/java/com/kape/payments/domain/GetSubscriptionsUseCase.kt
@@ -1,6 +1,6 @@
 package com.kape.payments.domain
 
-import com.kape.payments.data.Subscription
+import com.kape.data.dip.Subscription
 import org.koin.core.annotation.Singleton
 
 @Singleton
