+++ core/payments/src/google/java/com/kape/payments/ui/DipSubscriptionPaymentProvider.kt
@@ -1,7 +1,7 @@
 package com.kape.payments.ui
 
 import android.app.Activity
-import com.kape.payments.data.DipPurchaseData
+import com.kape.data.dip.DipPurchaseData
 
 interface DipSubscriptionPaymentProvider {
     /**
