+++ core/payments/src/google/java/com/kape/payments/ui/VpnSubscriptionPaymentProvider.kt
@@ -1,7 +1,7 @@
 package com.kape.payments.ui
 
 import android.app.Activity
-import com.kape.payments.data.Subscription
+import com.kape.data.dip.Subscription
 import com.kape.payments.data.SubscriptionPlan
 import com.kape.payments.utils.PurchaseHistoryState
 import com.kape.payments.utils.PurchaseState
