+++ core/payments/src/testGoogle/java/com/kape/payments/data/SubscriptionDataSourceImplTest.kt
@@ -1,5 +1,6 @@
 package com.kape.payments.data
 
+import com.kape.data.dip.Subscription
 import com.kape.payments.domain.SubscriptionDataSource
 import com.kape.payments.prefs.SubscriptionPrefs
 import com.privateinternetaccess.account.AccountRequestError
