+++ core/payments/src/testGoogle/java/com/kape/payments/domain/GetSubscriptionsUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.kape.payments.domain
 
-import com.kape.payments.data.Subscription
+import com.kape.data.dip.Subscription
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
