+++ features/dedicatedip/src/main/java/com/kape/dedicatedip/data/DipDataSourceImpl.kt
@@ -1,16 +1,11 @@
 package com.kape.dedicatedip.data
 
-import android.content.Context
 import com.kape.data.DI
 import com.kape.dedicatedip.domain.DipDataSource
 import com.kape.dedicatedip.utils.DipApiResult
 import com.kape.localprefs.prefs.DipPrefs
-import com.kape.payments.data.DipPurchaseData
 import com.privateinternetaccess.account.AndroidAccountAPI
-import com.privateinternetaccess.account.model.request.AndroidAddonSignupInformation
-import com.privateinternetaccess.account.model.response.AndroidAddonsSubscriptionsInformation
 import com.privateinternetaccess.account.model.response.DedicatedIPInformationResponse
-import com.privateinternetaccess.account.model.response.DipCountriesResponse
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.suspendCancellableCoroutine
@@ -20,15 +15,10 @@ import kotlin.coroutines.resume
 
 @Singleton([DipDataSource::class])
 class DipDataSourceImpl(
-    private val context: Context,
     private val accountApi: AndroidAccountAPI,
     private val dipPrefs: DipPrefs,
     @Named(DI.IO_SCOPE) private val ioScope: CoroutineScope,
 ) : DipDataSource {
-    companion object {
-        private const val STORE = "google_play"
-    }
-
     override suspend fun activate(ipToken: String): DipApiResult =
         suspendCancellableCoroutine { cont ->
             accountApi.redeemDedicatedIPs(listOf(ipToken)) { details, errors ->
@@ -61,79 +51,6 @@ class DipDataSourceImpl(
                 } ?: run {
                     cont.resume(DipApiResult.Error)
                 }
-            }
-        }
-
-    override suspend fun renew(ipToken: String): DipApiResult =
-        suspendCancellableCoroutine { cont ->
-            accountApi.renewDedicatedIP(ipToken) {
-                if (it.isNotEmpty()) {
-                    cont.resume(DipApiResult.Error)
-                    return@renewDedicatedIP
-                }
-                // todo: update if needed once implemented
-                cont.resume(DipApiResult.Active)
-            }
-        }
-
-    override suspend fun supportedCountries(): DipCountriesResponse? =
-        suspendCancellableCoroutine { cont ->
-            accountApi.supportedDedicatedIPCountries { details, errors ->
-                if (errors.isNotEmpty()) {
-                    cont.resume(null)
-                    return@supportedDedicatedIPCountries
-                }
-                cont.resume(details)
-            }
-        }
-
-    override suspend fun signupPlans(): AndroidAddonsSubscriptionsInformation? =
-        suspendCancellableCoroutine { cont ->
-            accountApi.addonsSubscriptions { details, errors ->
-                if (errors.isNotEmpty()) {
-                    cont.resume(null)
-                    return@addonsSubscriptions
-                }
-                cont.resume(details)
-            }
-        }
-
-    override suspend fun signup(dipPurchaseData: DipPurchaseData): Result<Unit> =
-        suspendCancellableCoroutine { cont ->
-            accountApi.addonSignUp(
-                AndroidAddonSignupInformation(
-                    receipt =
-                        AndroidAddonSignupInformation.Receipt(
-                            applicationPackage = context.packageName,
-                            productId = dipPurchaseData.productId,
-                            orderId = dipPurchaseData.orderId,
-                            token = dipPurchaseData.token,
-                        ),
-                    store = STORE,
-                ),
-            ) { errors ->
-                if (errors.isNotEmpty()) {
-                    cont.resume(Result.failure(IllegalStateException("Signup validation failed")))
-                    return@addonSignUp
-                }
-                cont.resume(Result.success(Unit))
-            }
-        }
-
-    override suspend fun fetchToken(
-        countryCode: String,
-        regionName: String,
-    ): Result<String> =
-        suspendCancellableCoroutine { cont ->
-            accountApi.getDedicatedIP(
-                countryCode = countryCode,
-                regionName = regionName,
-            ) { details, errors ->
-                if (details == null || errors.isNotEmpty()) {
-                    cont.resume(Result.failure(IllegalStateException("Fetch token failed")))
-                    return@getDedicatedIP
-                }
-                cont.resume(Result.success(details.token))
             }
         }
 }
