+++ features/dedicatedip/src/main/java/com/kape/dedicatedip/di/DipModule.kt
@@ -1,27 +1,16 @@
 package com.kape.dedicatedip.di
 
-import android.content.Context
 import com.kape.buildconfig.data.BuildConfigProvider
 import com.kape.contracts.ConnectionManager
 import com.kape.contracts.Router
 import com.kape.data.DI
 import com.kape.dedicatedip.data.DipDataSourceImpl
-import com.kape.dedicatedip.data.DipSignupRepository
 import com.kape.dedicatedip.domain.ActivateDipUseCase
 import com.kape.dedicatedip.domain.DipDataSource
-import com.kape.dedicatedip.domain.FetchSignupDipToken
-import com.kape.dedicatedip.domain.GetDipMonthlyPlan
-import com.kape.dedicatedip.domain.GetDipSupportedCountries
-import com.kape.dedicatedip.domain.GetDipYearlyPlan
-import com.kape.dedicatedip.domain.RenewDipUseCase
-import com.kape.dedicatedip.domain.ValidateDipSignup
+import com.kape.dedicatedip.domain.DipPurchaseHandler
 import com.kape.dedicatedip.ui.vm.DipViewModel
 import com.kape.localprefs.prefs.ConnectionPrefs
 import com.kape.localprefs.prefs.DipPrefs
-import com.kape.payments.prefs.SubscriptionPrefs
-import com.kape.payments.ui.DipSubscriptionPaymentProvider
-import com.kape.payments.ui.VpnSubscriptionPaymentProvider
-import com.kape.ui.utils.PriceFormatter
 import com.kape.vpnregions.utils.RegionListProvider
 import com.privateinternetaccess.account.AndroidAccountAPI
 import kotlinx.coroutines.CoroutineDispatcher
@@ -35,84 +24,35 @@ import org.koin.core.annotation.Singleton
 class DipModule {
     @Singleton(binds = [DipDataSource::class])
     fun provideDipDataSource(
-        context: Context,
         accountApi: AndroidAccountAPI,
         dipPrefs: DipPrefs,
         @Named(DI.IO_SCOPE) ioScope: CoroutineScope,
-    ): DipDataSource = DipDataSourceImpl(context, accountApi, dipPrefs, ioScope)
-
-    @Singleton
-    fun provideDipSignupRepository(
-        dipDataSource: DipDataSource,
-        dipPrefs: DipPrefs,
-    ): DipSignupRepository = DipSignupRepository(dipDataSource, dipPrefs)
+    ): DipDataSource = DipDataSourceImpl(accountApi, dipPrefs, ioScope)
 
     @Singleton
     fun provideActivateDipUseCase(dataSource: DipDataSource): ActivateDipUseCase = ActivateDipUseCase(dataSource)
-
-    @Singleton
-    fun provideRenewDipUseCase(dataSource: DipDataSource): RenewDipUseCase = RenewDipUseCase(dataSource)
-
-    @Singleton
-    fun provideValidateDipSignup(
-        subscriptionPrefs: SubscriptionPrefs,
-        dataSource: DipDataSource,
-    ): ValidateDipSignup = ValidateDipSignup(subscriptionPrefs, dataSource)
-
-    @Singleton
-    fun provideGetDipSupportedCountries(dipSignupRepository: DipSignupRepository): GetDipSupportedCountries =
-        GetDipSupportedCountries(dipSignupRepository)
-
-    @Singleton
-    fun provideGetDipMonthlyPlan(
-        dipSignupRepository: DipSignupRepository,
-        dipSubscriptionPaymentProvider: DipSubscriptionPaymentProvider,
-        formatter: PriceFormatter,
-    ): GetDipMonthlyPlan = GetDipMonthlyPlan(dipSignupRepository, dipSubscriptionPaymentProvider, formatter)
-
-    @Singleton
-    fun provideGetDipYearlyPlan(
-        dipSignupRepository: DipSignupRepository,
-        dipSubscriptionPaymentProvider: DipSubscriptionPaymentProvider,
-        formatter: PriceFormatter,
-    ): GetDipYearlyPlan = GetDipYearlyPlan(dipSignupRepository, dipSubscriptionPaymentProvider, formatter)
-
-    @Singleton
-    fun provideFetchSignupDipToken(dipDataSource: DipDataSource): FetchSignupDipToken = FetchSignupDipToken(dipDataSource)
 
     @KoinViewModel
     fun provideDipViewModel(
         router: Router,
         regionListProvider: RegionListProvider,
         activateDipUseCase: ActivateDipUseCase,
-        getDipSupportedCountries: GetDipSupportedCountries,
-        getDipMonthlyPlan: GetDipMonthlyPlan,
-        getDipYearlyPlan: GetDipYearlyPlan,
-        validateDipSignup: ValidateDipSignup,
-        fetchSignupDipToken: FetchSignupDipToken,
-        vpnSubscriptionPaymentProvider: VpnSubscriptionPaymentProvider,
-        dipSubscriptionPaymentProvider: DipSubscriptionPaymentProvider,
         dipPrefs: DipPrefs,
         connectionPrefs: ConnectionPrefs,
         buildConfigProvider: BuildConfigProvider,
         connectionManager: ConnectionManager,
+        dipPurchaseHandler: DipPurchaseHandler,
         @Named(DI.IO_DISPATCHER) ioDispatcher: CoroutineDispatcher,
     ): DipViewModel =
         DipViewModel(
             router,
             regionListProvider,
             activateDipUseCase,
-            getDipSupportedCountries,
-            getDipMonthlyPlan,
-            getDipYearlyPlan,
-            validateDipSignup,
-            fetchSignupDipToken,
-            vpnSubscriptionPaymentProvider,
-            dipSubscriptionPaymentProvider,
             dipPrefs,
             connectionPrefs,
             buildConfigProvider,
             connectionManager,
+            dipPurchaseHandler,
             ioDispatcher,
         )
 }
