--- features/signup/src/noinapp/java/com/kape/signup/domain/SignupUseCase.kt
@@ -1,17 +1,0 @@
-package com.kape.signup.domain
-
-import com.kape.login.domain.mobile.LoginUseCase
-import com.kape.payments.domain.GetPurchaseDetailsUseCase
-import com.kape.signup.data.models.Credentials
-import org.koin.core.annotation.Singleton
-
-@Singleton
-class SignupUseCase(
-    private val signupDataSource: SignupDataSource,
-    private val loginUseCase: LoginUseCase,
-    private val emailDataSource: EmailDataSource,
-    private val purchaseDetailsUseCase: GetPurchaseDetailsUseCase,
-    private val getObfuscatedDeviceIdentifierUseCase: GetObfuscatedDeviceIdentifierUseCase,
-) {
-    suspend fun vpnSignup(email: String): Credentials? = null
-}
